initial but massive step towards a stable abstract model
OK, this is pretty big. Let's explain things a bit:

modules as packages
-------------------

The idea is that every package under com.bio4j.model corresponds to a
module. There are probably Nodes and Rels out of place here, but I
think that is more or less OK overall.

add properties and property types
----------------------------------

I created a few of them, they are under the

- com.bio4j.model.properties

package. I also used this in some nodes as a sample:

- com.bio4j.model.isoforms.nodes.Isoform

I tried all sort of alternatives here, and I think this is by far the
most convenient I could find.